New: the ability to specify a revision range for the non-tag based analysis, if no range is specified the entire commit history of a given file is implied
Modified:
when specifying a revision range it turns out that some commits
captured during the git blame queries are not capture by the git log
query. I added a second git log query to capture all commits for the
file. This is inefficient but there doesn't seem to be an obvious
better solution at the moment. 


Signed-off-by: Mitchell Joblin <joblin.m@gmail.com>