Make dispatcher version agnostic
Version numbering differs considerably from project to project,
so do not hard-code any assumptions into the dispatcher. Pass
the desired revision boundaries as mandatory parameters to the
shell script.

Note that this removes the ability to iterate over multiple revision
ranges, that is, prepare multiple reports. However, this can easily
be recovered with another dispatching stage.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>