Define a yaml configuration file format to describe projects
This eliminates the need to pass all parameters via command
line options, which is cumbersome. Simple parsers are provided
for R and python.

The configuration file specifies the local name under which the
analysis is stored in the file system (project), the identifier
of the git repository (repo), a human readable textual identifier
for the project (description -- use only a few words), and
the revision path that should be traversed. Release candidate
revisions that map to the revisions are specified in rcs. This
element may be omitted if the project does not use release candidates.
Versions for which no rc exists are specified with an empty string ("").

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>