Silence the time series analysis
Status output that a date has been parsed is not that interesting,
so avoid this. Additionally, sapply is used to create the final
set of plots, but the result is a side effect, which generates
a list of NULL elements. Assign this to a dummy variable to avoid
the result being printed to the console.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>