Added function to randomize the commit orginization withing a file
The idea behind this was, the way developers make contributions to
a file holds some kind of organization. For example, developers
who are working together towards a common goal will  more often
make contributions in close proximity (similar location in the file).
This organization leads to the strong community structure seen in
the collaboration graphs. This function randomizes the commit
organization to see if this destroys the strength of community
structure seen compared to the non-randomized collaboration graphs.

Signed-off-by: Mitchell Joblin <joblin.m@gmail.com>