Add function to compute the collaboration network using author and committer
The function creates the collaboration netword by introducing a one
directional relationship from the individual who performs the commit
(ie. committer) towards the person who authored the commit. The committer
is aware of the authors contribution and therefore is collaborating to
some degree. However the author is not necessarily aware of the committer
and his/her intents, thus the one directional relationship.

Signed-off-by: Mitchell Joblin <joblin.m@gmail.com>