Bring the different network building methods together
The various methods for building the contribution networks
(tag based, proximity based etc.) had diverged even in areas
were they were similar. The personInfo class was also very
tag-centric (functions and naming scheme) this was addressed
and the personInfo class in now more generic in terms of handling
"links" not just tags. This way we can provide a more uniform
interface for the differing network construction methods. At
this point only the proximity based method is not fully integrated
into the modifications made to personInfo.

Signed-off-by: Mitchell Joblin <joblin.m@gmail.com>