added function
The new function allows one to compute the largest connected
subgraph and return the igraph object. The previous method
was to use compute.largest.subgraph(..) which then returns
the indices of component. A further step of building a new
igraph object from these indices was annoying and the indices
are not allways needed.

Signed-off-by: Mitchell Joblin <joblin.m@gmail.com>