added functions to perform a significance test for community structure
The functionality is not fully implemented yet. The goal is to develop
a strategy to compute the statistical significance of a community
structure compared to a random graph with similar properties. The
current implementation committed here uses a rewiring concept where
the random graph used for comparison has its degree maintained after
the randomization process. The randomization is computed many times
to compute a distribution of conductance for each random graph clustering
we would then use this distribution to see if the conductance measured
in the original graph is statistically significant.

Signed-off-by: Mitchell Joblin <joblin.m@gmail.com>