generalized the _gitQuery(..) function
The _gitQuery(..) function was used to submit git commands
to the system terminal. Thefunction is  more general than for use with
git commands since it only submits a string to the system
command line. The change reflects this generality to avoid
confusion when using the function in non-git based commands.

Signed-off-by: Mitchell Joblin <joblin.m@gmail.com>