igraph: Always work with weighted matrices
The adjacency matrices are weighted for all three analysis
methods. Previous bugs in igraph have prevented us from using a
weighted analysis for the tagged case, but since this is fixed
upstream, we can now use a weighted decomposition in any case.
We can thus get rid of the weighted argument to performGraphAnalysis().

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>