Construct release ranges in the dispatcher
From the date in the configuration file, it is
perfectly possible to construct the release ranges although
we don't have any specific dates yet, so fill in the
information early.

Since we need IDs for RC tags, also enter them into the database.
This requires some changes to when the actual dates are filled
in during later analysis stages.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>