Fix rc date handling in the db access layer
ymd_hms() has an odd convention: When a vector of NAs is
passed, the function boils out with an error. When one single
entry is correct, all other NSa are happily ignored and
converted to NAs.

The previous check that tried to avoid the failure case was
wrong; it can well be possible that the first release range
does not include an RC cycle, whereas others do. We need
to check if really all vector elements are NAs, and in this
case skip the conversion.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>