Make get.plot.id() usable without conf object
When multiple projects will be compared with each other during
dynamic visualisation, we will not have a configuration object
for each of the projects. Nonetheless, we need to be able
to query plot IDs, so provide a new helper function with a
different signature. Implement the previous function on top of the
new function.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>