Write edge lists for all clusters into the database
For each cluster that is written into the data base,
store a simple representation that contains only the
involved contributors and their page ranks. Additionally,
the edge list (in the form of weighted edges) is stored
in another table. This allows us to faithfully reconstruct
all clusters in the analysis phases outside person.r, and
also construct time series over the cluster behaviour.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>