Add helper function to obtain in-DB pagerank id
Entries in table pagerank_matrix are "guarded" by an entry
in table pagerank that orders page rank value by release range
and clustering technique. Provide a helper function to
get the ID for a given combination of range/technique,
possibly creating a new ID if it does not exist yet.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>