Avoid code duplication when saving cluster graphs
The decomposition into communities is independent of whether
transposed or regular page rank is used. While we need to generate
different dot files for each combination, it suffices to write
the clusters (vertices and edge lists) into the data base once.
Refactor the code to do this without duplication. And add
some documentation to clarify the situation.

Since the structure of the generated dot file names changes,
the report generator needs to be updated, too.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>