Community selection: Provide a reasonable heuristic
The small communities produces by some clusterings require
that too small contributions are removed (for instance, 1 and
two person communities). However, using a fixed threshold is
not uncontroversial because depending on the size of a project,
a community with three contributors might be noise or utterly
essential.

The new selection function takes a minimum fraction of contributors
that is supposed to remain in the set (by default 95%), and
picks which communities to remove by this criterion. It is also
possible to specify an upper bound on the size of removed communities.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>