Dynamically visualise mailing list time series
Add a shiny script that visualises the time series associated
with mailing lists, and also allows for performing various
simple transformations that can be done at run-time.

NOTE: The example is not yet complete because some infrastucture
is missing in the database to determine which mailing lists belong
to which project; currently, hard-coded values for project and
a hard-coded mailing list name are used. The code is checked in
early to facilitate coupling it with the web frontend.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>