Add file with clustering definitions
We support multiple clustering types, and we also support different
page rank calculation methods. Add a support file to provide central
definitions of the associated (string and integer) constants,
and two helper functions to check if a given method/type is valid.

Adapt the existing code to use these definitions.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>