Compute per-cluster statistics and plots based on the database content
Redesign the calculations that were based on *_cluster_stats.txt to
solely rely on data available in the data base, which allows us
to get rid of another set of files that are stored on the file
system. Consequently, we can also get rid of function
gen.cluster.file.list().

This commit finally eliminates the need for temporary files, and
makes the analysis completely DB based.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>