Commit time series: Properly handle rc/non-RC and tagged/non-tagged cases
Whether a project uses tags in a release cycle or not is completely
independent of the question whether RC candidate phases are used
or not. Nonetheless, the items were improperly entangled in the code
so far, leading to errors in some combinations. Carefully distinguish
between the possible variants, and prepare the correct plots in all
cases.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>