Update include paths in the mailing list analysis
With moving the code into the main prosoda repository,
the include paths have changed -- but they are now guaranteed
to be stable and don't require any assumptions on where
the two separate repositories are checked out.
In fact, since batch.r is supposed to be called from the
main repo path, we need to treat all include files local
to the main repo directory, which makes it possible to use
the same setting in the main prosoda code and the mailing list
analysis.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>