ML analysis: Warn if a http proxy is used
The mailing list analysis code communicated with the ID
service via a http REST interface -- usually, the services
run on the same hosts, or in neighbouring machines that
communicate directly with each other. Having an active http
proxy is usually not what is intended in this scenario, so
issue a warning if one is detected.

Since there could be use cases in which a proxy may make sense,
continue the operation nevertheless, and don't abort the analysis.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>