Determine ID service hostname/port from the global configuration
Said data is not a project-specific property, but a global
one. Consequently, use the global configuration file to
store and validate the information.
However, store the information also in the local configuration
file once it is available; this makes it easier to pass it
around between functions that anyway exchange conf objects/tokens.

Reported-by: Johannes Ebke <johannes.ebke.ext@siemens.com>
Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>