Move all code into the prosoda python package
To enable simple packaging, all python, R and perl code is moved into
the prosoda python package. All python-to-python calls are now in the
same interpreter, and R and perl file discovery use the pkg_resources
mechanism. This changeset should not change the functionality, and it
has been checked that the output of a test run is virtually identical
to the pre-refactoring version.

Signed-off-by: Johannes Ebke <johannes.ebke.ext@siemens.com>
Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>