Change spin glass community detection algorithm parameter
The spin glass algorithm uses a probability distribution
of edges to determine the expected number of edges that
exist in a random graph. Two randomization models are
possible, one in which the degree distribution of nodes
is maintained namely the configuation model. This change
forces the spin glass algorithm to use the configuation
model which is known to be more appropriate for our
purposes (ie. social graph like situations).

Signed-off-by: Mitchell Joblin <joblin.m@gmail.com>
Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>