Handle ML command line option for the number of cores
Fix how the number of cores is determined from the command
line options. Also switch from using "nodes" to "cores"; the
former terminology is a left-over from earlier versions of the
code that relied on MPI for parallelisation. Since we use a fork
model now, the name is not only inaccurate, but downright misleading.