Do not clobber database on a prosoda rerun
Before new data is added to a table, the previous version of the same
plot, mailinglist, cluster or pagerank is deleted. To keep it this way,
any id getters in db.r should now be available in a "clearing" version
get.clear.id and a non-clearing version get.id for reading. Fixes #67.

Signed-off-by: Johannes Ebke <johannes.ebke.ext@siemens.com>
Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>