Improve interval performance by sorting date list
Until now, finding the number of messages in a given interval was done
by checking every message date against the interval. This lead to
multi-minute wait times for e.g. removing empty intervals, since many
intervals are checked against O(500000) messages. Now the dates list is
sorted first, and the interval functions are amended to expect a sorted
list of dates as input.

Signed-off-by: Johannes Ebke <johannes.ebke.ext@siemens.com>
Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>