Add shiny-server configuration and install docs
Add a working shiny server configuration, run and installation scripts,
as well as documented the installation in the INSTALL file. As an
example, the dynamic graph "contributors" has been converted to work
with the shiny server. This currently breaks the use as a standalone
dynamic graph, but this is not expected to be a permanent state.

Signed-off-by: Johannes Ebke <johannes.ebke.ext@siemens.com>
Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>