Performance improvements and getting the project ID by URL
In the ml.timeseries.server function, loading of the timeseries data is done
outside the reactive environment in, so it will be done only once.
The reactive input from the session variable is used to extract the
URL parameters (using shiny's parseQueryString function).
Isolation of the session variable allows to extract the values to outside
of the reactive environment. The Project ID from URL parameter "projectid"
is used to select the data for this server.
Some checking of the Project ID "pid" is done and execution is stoped in case of
nonexisting parameter or parameter value.
Execution will also stop, if the pid does not exist in the database (see db.r)

Signed-off-by: Albert Eckert <albert.eckert@siemens.com>
Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>