Fix incorrect dist function in commit.structure
Due to the new structure, ts_utils.r is sourced when commit.structure is
executed. In ts_utils.r, the dtw library is included, which pulls in the
proxy library. The proxy library overwrites the "dist" function of the
stats package which is used in commit.structure. This patch changes
calls of dist explicitely to stats::dist to fix this issue.

Signed-off-by: Johannes Ebke <johannes.ebke.ext@siemens.com>