Reworked multiprocessing Job Pool to solve #81
Refactored the BatchJob object into a per-instance BatchJobPool. To
solve the sporadic stuck process problem, do not use threads in the
Process management anymore, as some interactions seemed to be
problematic. Also use a fixed worker pool, and terminate immediately if
one of the workers dies. Finally, add some tests for the new pool.

Signed-off-by: Johannes Ebke <johannes.ebke.ext@siemens.com>