Workaround for numerical stability problems with page.rank
igraph's page rank implementation sometimes fails to correctly
compute some low-level matrix decomposition. The corner case
when this happens is not completely understood, but it is rooted
deep down in some ARPACK fortran code.

For now, we can possibly solve the problem by trying to compute the page
rank using a different random seed, and re-trying this until the
decomposition succeeds, with a limit of 10 on the number of
retries. Using new copies of the graph also seems to help.
Admittedly, it's unclear if this removes all symptoms, or makes
them just more unlikely. The whole commit is therefore officially
classified as black magi.

This is intended to go away after the root cause has been
determined and _properly_ fixed on the fortran level.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>