Refactor widgets to use reactive values
Up to now, widgets were regenerated when reactive configuration values
changed. This leads to the effect that the "fadeout" when updating pages
does not work anymore, and an important hint that computation is ongoing
is missing. Also, this refactoring enables easiert reuse of computed
data in different sub-methods of a widget.

Signed-off-by: Johannes Ebke <johannes.ebke.ext@siemens.com>