Fix time stamp analysis for single commits
When the time stamp of a single commit is required, we must not
sort the list of commits that is determined by getCommitIDsLL().
This can bring commits other than the one originally intended to
the top of the list, resulting in bogus dates (alternatively,
we could also use git log -1 to obtain only the commit in question).

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>