Establish per-thread database connections
Multiple threads are used in the mailing list analysis, but the
connection to the mysql database is established before the threads
are created -- it's thus shared among all worker threads. This
can lead to various problems, see http://TODO.

Close the connection before creating the threads, and establish
a separate connection for each worker.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>