Refactor observe/reactive contexts in dashboard server
To increase the stability and realiability of the dashboard server, the
observe/reactive contexts in server.r were refactored into smaller pieces.
It remains to be done to update the widget help text and colour via
special messages, the widget probably has to be recreated on the client
side for the latter.

Signed-off-by: Albert Eckert <albert.eckert@siemens.com>
Signed-off-by: Johannes Ebke <johannes.ebke.ext@siemens.com>